// src/generated/model/adresserFrontendValg.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type AdresserFrontendValg = (typeof AdresserFrontendValg)[keyof typeof AdresserFrontendValg];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AdresserFrontendValg = {
    folkeregistrert: "folkeregistrert",
    midlertidig: "midlertidig",
    soknad: "soknad",
} as const;